Fix metdata password for cluster and containers
There were three ways the deafult for this parameter was defined:
    cluster:
        password: metadataPass
    container:
        password: ${_param:nova_metadata_password}
    single:
        password: ${_param:metadata_password}

Between "nova_metadata_password" and "metadata_password",
"metadata_password" seems better because this password is intentionally
shared between nova and neturon, and it should not be different between
the two.  Standardize all three service roles to use the same default.